+++ parsley-core/src/ghc-8.10+/Parsley/Internal/Backend/Machine/Types/Input/Offset.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE DerivingStrategies, UnboxedTuples #-}
 {-|
-Module      : Parsley.Internal.Backend.Machine.Types.Offset
+Module      : Parsley.Internal.Backend.Machine.Types.Input.Offset
 Description : Statically refined offsets.
 License     : BSD-3-Clause
 Maintainer  : Jamie Willis
@@ -10,7 +10,7 @@ This module contains the statically refined `Offset` type,
 which can be used to reason about input in different parts of
 a parser as it is evaluated.
 
-@since 1.4.0.0
+@since 1.8.0.0
 -}
 module Parsley.Internal.Backend.Machine.Types.Input.Offset (
     Offset, mkOffset, offset, moveOne, moveN, same,